remove superflous 4.6 dep
git-svn-id: https://cloudcity.svn.sourceforge.net/svnroot/cloudcity@1102 d23db356-9f35-0410-88be-f4158e15d4e5

+++ XBar/xbar.cpp
@@ -143,7 +143,7 @@ XBar::init()
     // TODO: use plasmoid popup and make this dynamic -> update all menubars...
     QSettings settings("Bespin", "XBar");
     settings.beginGroup("XBar");
-    float scale = settings.value("FontScale", 1.0f).toFloat();
+    double scale = settings.value("FontScale", 1.0f).toDouble();
     if (scale > 0.0 && scale != 1.0)
     {
         myFont = KGlobalSettings::menuFont();
